refactor: move matcher test where other Sample types can be checked
