+++ internal/agent/agent_test.go
@@ -915,86 +915,6 @@ func Test_ProcessSampling(t *testing.T) {
 	}
 }
 
-func Test_ProcessSamplingExcludes(t *testing.T) {
-	t.Parallel()
-
-	someSample := &types.ProcessSample{
-		ProcessDisplayName: "some-process",
-	}
-
-	boolAsPointer := func(val bool) *bool {
-		return &val
-	}
-
-	type testCase struct {
-		name          string
-		c             *config.Config
-		ff            feature_flags.Retriever
-		expectInclude bool
-	}
-	testCases := []testCase{
-		{
-			name:          "Include not matching must not include",
-			c:             &config.Config{EnableProcessMetrics: boolAsPointer(true), IncludeMetricsMatchers: map[string][]string{"process.name": {"does-not-match"}}, DisableCloudMetadata: true},
-			ff:            test.NewFFRetrieverReturning(false, false),
-			expectInclude: false,
-		},
-		{
-			name:          "Include matching should not exclude",
-			c:             &config.Config{EnableProcessMetrics: boolAsPointer(true), IncludeMetricsMatchers: map[string][]string{"process.name": {"some-process"}}, DisableCloudMetadata: true},
-			ff:            test.NewFFRetrieverReturning(false, false),
-			expectInclude: true,
-		},
-		{
-			name:          "Exclude matching should exclude with process metrics enabled",
-			c:             &config.Config{EnableProcessMetrics: boolAsPointer(true), ExcludeMetricsMatchers: map[string][]string{"process.name": {"some-process"}}, DisableCloudMetadata: true},
-			ff:            test.NewFFRetrieverReturning(false, false),
-			expectInclude: false,
-		},
-		{
-			name:          "Exclude matching should exclude",
-			c:             &config.Config{ExcludeMetricsMatchers: map[string][]string{"process.name": {"some-process"}}, DisableCloudMetadata: true},
-			ff:            test.NewFFRetrieverReturning(false, false),
-			expectInclude: false,
-		},
-		{
-			name:          "Exclude not matching should not exclude with process metrics enabled",
-			c:             &config.Config{EnableProcessMetrics: boolAsPointer(true), ExcludeMetricsMatchers: map[string][]string{"process.name": {"does-not-match"}}, DisableCloudMetadata: true},
-			ff:            test.NewFFRetrieverReturning(false, false),
-			expectInclude: true,
-		},
-		{
-			name:          "Exclude not matching should not exclude",
-			c:             &config.Config{ExcludeMetricsMatchers: map[string][]string{"process.name": {"does-not-match"}}, DisableCloudMetadata: true},
-			ff:            test.NewFFRetrieverReturning(false, false),
-			expectInclude: false,
-		},
-		{
-			name:          "Include matching should include even if exclude is configured with process metrics enabled",
-			c:             &config.Config{EnableProcessMetrics: boolAsPointer(true), IncludeMetricsMatchers: map[string][]string{"process.name": {"some-process"}}, ExcludeMetricsMatchers: map[string][]string{"process.name": {"some-process"}}, DisableCloudMetadata: true},
-			ff:            test.NewFFRetrieverReturning(false, false),
-			expectInclude: true,
-		},
-		{
-			name:          "Include matching should be include even when enable_process_metrics is not defined (nil)",
-			c:             &config.Config{IncludeMetricsMatchers: map[string][]string{"process.name": {"some-process"}}, ExcludeMetricsMatchers: map[string][]string{"process.name": {"some-process"}}, DisableCloudMetadata: true},
-			ff:            test.NewFFRetrieverReturning(false, false),
-			expectInclude: true,
-		},
-	}
-
-	for _, tc := range testCases {
-		testCase := tc
-		a, _ := NewAgent(testCase.c, "test", "userAgent", testCase.ff)
-
-		t.Run(testCase.name, func(t *testing.T) {
-			t.Parallel()
-
-			assert.Equal(t, testCase.expectInclude, a.Context.includeEvent(someSample))
-		})
-	}
-}
-
 func Test_ProcessSamplingExcludesAllCases(t *testing.T) {
 	t.Parallel()
 
@@ -1188,7 +1108,7 @@ func Test_ProcessSamplingExcludesAllCases(t *testing.T) {
 		t.Run(testCase.name, func(t *testing.T) {
 			t.Parallel()
 
-			assert.Equal(t, testCase.expectInclude, a.Context.includeEvent(someSample))
+			assert.Equal(t, testCase.expectInclude, a.Context.IncludeEvent(someSample))
 		})
 	}
 }
